rename SchemaNewAdmin to SchemaAdmin
